Moved ci-tests script to top-level of repo

+++ ci-tests.bash
@@ -1,8 +1,10 @@
 #!/bin/bash
 set -e
-trap 'mv -v /opt/clean/etc/IDEEnvs{.bak,}' EXIT
-cp -v /opt/clean/etc/IDEEnvs{,.bak}
-sed -i "s|{Application}/lib/iTasks|$(pwd)/Libraries|g" /opt/clean/etc/IDEEnvs
+if [ -e /opt/clean/etc/IDEEnvs ]; then
+	trap 'mv -v /opt/clean/etc/IDEEnvs{.bak,}' EXIT
+	cp -v /opt/clean/etc/IDEEnvs{,.bak}
+	sed -i "s|{Application}/lib/iTasks|$(pwd)/Libraries|g" /opt/clean/etc/IDEEnvs
+fi
 
 #Try to compile everything
 find . -name "*.prj.default" | while read f; do
@@ -11,4 +13,4 @@ find . -name "*.prj.default" | while read f; do
 find . -name "*.prj" -exec dirname {} \; | sort -u | xargs -I{} sh -c "cd {}; cpm make"
 
 #Run the unit tests
-find Unit -type f -perm +111 -exec cleantest -r {} \;
+find Tests/Unit -type f -perm +111 -exec cleantest -r {} \;
